Move config run code out into it's own file
